day10_05

+++ src/lambda/ex03/Calculable.java
@@ -3,5 +3,5 @@ package lambda.ex03;
 @FunctionalInterface
 public interface Calculable {
     //추상 메소드
-    double calculate(int x,int y);
+    double calculate(double x,double y);
 }
